Convert radio buttons to component file
